+++ CHANGELOG.md
@@ -7,7 +7,8 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Bug Fixes
 
--
+- Fix multidimensional array encoding support for 2D arrays [#2287](https://github.com/LFDT-web3j/web3j/pull/2287)
+
 
 ### Features